Restructure the larky project. Fixes #18.
