+++ larky/src/main/java/com/verygood/security/larky/modules/C99MathModule.java
@@ -1,4 +1,4 @@
-package com.verygood.security.larky.nativelib.std;
+package com.verygood.security.larky.modules;
 
 import com.google.common.math.DoubleMath;
 
@@ -14,9 +14,9 @@ import net.starlark.java.eval.StarlarkValue;
     name = "c99math",
     category = "BUILTIN",
     doc = "This module provides access to the mathematical functions defined by the C99 standard")
-public class C99Math implements StarlarkValue {
+public class C99MathModule implements StarlarkValue {
 
-  public static final C99Math INSTANCE = new C99Math();
+  public static final C99MathModule INSTANCE = new C99MathModule();
 
   @StarlarkMethod(name = "PI", doc = "a constant pi", structField = true)
   public StarlarkFloat PI_CONSTANT() {
