+++ larky/src/main/java/com/verygood/security/larky/modules/globals/PythonBuiltins.java
@@ -1,4 +1,4 @@
-package com.verygood.security.larky.nativelib;
+package com.verygood.security.larky.modules.globals;
 
 import com.verygood.security.larky.annot.Library;
