+++ larky/src/main/java/com/verygood/security/larky/modules/testing/README.md
@@ -1,0 +1,5 @@
+# Testing Modules 
+
+These modules expose testing native modules that are **NOT SAFE** to use in normal Larky embeds. 
+
+As a result, this namespace is only exposed via the LarkyTest suites.
