+++ larky/src/main/java/com/verygood/security/larky/modules/types/structs/SimpleStruct.java
@@ -1,15 +1,12 @@
-package com.verygood.security.larky.stdtypes.structs;
+package com.verygood.security.larky.modules.types.structs;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 
-import net.starlark.java.annot.StarlarkMethod;
 import net.starlark.java.eval.Dict;
 import net.starlark.java.eval.Printer;
 import net.starlark.java.eval.Starlark;
-import net.starlark.java.eval.StarlarkCallable;
 import net.starlark.java.eval.StarlarkSemantics;
-import net.starlark.java.eval.StarlarkThread;
 import net.starlark.java.eval.Structure;
 
 import java.util.Map;
@@ -74,57 +71,6 @@ public class SimpleStruct implements Structure {
       sep = ", ";
     }
     p.append(")");
-  }
-
-  // SimpleStructWithMethods augments SimpleStruct's fields with annotated Java methods.
-  private static final class SimpleStructWithMethods extends SimpleStruct {
-
-    // A function that returns "fromValues".
-    private static final Object returnFromValues =
-        new StarlarkCallable() {
-          @Override
-          public String getName() {
-            return "returnFromValues";
-          }
-
-          @Override
-          public Object fastcall(StarlarkThread thread, Object[] positional, Object[] named) {
-            return "bar";
-          }
-        };
-
-    SimpleStructWithMethods() {
-      super(
-          ImmutableMap.of(
-              "values_only_field",
-              "fromValues",
-              "values_only_method",
-              returnFromValues,
-              "collision_field",
-              "fromValues",
-              "collision_method",
-              returnFromValues));
-    }
-
-    @StarlarkMethod(name = "callable_only_field", documented = false, structField = true)
-    public String getCallableOnlyField() {
-      return "fromStarlarkMethod";
-    }
-
-    @StarlarkMethod(name = "callable_only_method", documented = false, structField = false)
-    public String getCallableOnlyMethod() {
-      return "fromStarlarkMethod";
-    }
-
-    @StarlarkMethod(name = "collision_field", documented = false, structField = true)
-    public String getCollisionField() {
-      return "fromStarlarkMethod";
-    }
-
-    @StarlarkMethod(name = "collision_method", documented = false, structField = false)
-    public String getCollisionMethod() {
-      return "fromStarlarkMethod";
-    }
   }
 
 }
