+++ larky/src/main/java/com/verygood/security/larky/utils/NullPrintStream.java
@@ -1,0 +1,12 @@
+package com.verygood.security.larky.utils;
+
+import com.google.common.io.ByteStreams;
+
+import java.io.PrintStream;
+
+public final class NullPrintStream extends PrintStream {
+  @SuppressWarnings("UnstableApiUsage")
+  public NullPrintStream() {
+    super(ByteStreams.nullOutputStream());
+  }
+}
