+++ larky/src/test/java/com/verygood/security/larky/LarkyTest.java
@@ -4,10 +4,10 @@ import static com.verygood.security.larky.ModuleSupplier.CORE_MODULES;
 
 import com.google.common.collect.ImmutableSet;
 
+import com.verygood.security.larky.modules.globals.LarkyGlobals;
+import com.verygood.security.larky.modules.globals.PythonBuiltins;
 import com.verygood.security.larky.console.testing.TestingConsole;
-import com.verygood.security.larky.nativelib.LarkyGlobals;
-import com.verygood.security.larky.nativelib.PythonBuiltins;
-import com.verygood.security.larky.nativelib.test.UnittestModule;
+import com.verygood.security.larky.modules.testing.UnittestModule;
 import com.verygood.security.larky.parser.LarkyScript;
 import com.verygood.security.larky.parser.ParsedStarFile;
 import com.verygood.security.larky.parser.PathBasedStarFile;
